# swagger.yml
swagger: "2.0"
info:
  description: "Responsible for taking Child Tax Credit eligibility requests and checking\
    \ them against HMRC data in an internal database before making a call to the real\
    \ HMRC API."
  title: "HMRC API Service"
  contact:
    name: "Department Of Health"
    url: "https://github.com/DepartmentOfHealth-htbhf"
    email: "dh-htbhf-team@equalexperts.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:40869"
basePath: "/"
tags:
- name: "hmrc-eligibility-controller"
  description: "Endpoints for dealing with HMRC Eligibility requests for Healthy Start."
paths:
  /v1/hmrc/eligibility:
    post:
      tags:
      - "hmrc-eligibility-controller"
      summary: "Retrieve the eligibility of a person for Healthy Start based on HMRC's\
        \ opinion of their income and the fact that they're getting Child Tax Credits"
      operationId: "getBenefitsUsingPOST"
      consumes:
      - "application/json"
      produces:
      - "*/*"
      parameters:
      - in: "body"
        name: "eligibilityRequest"
        description: "The eligibility request for HMRC for Healthy Start"
        required: false
        schema:
          $ref: "#/definitions/EligibilityRequest"
      responses:
        200:
          description: "The person's eligibility for Healthy Start from HMRC's point\
            \ of view"
          schema:
            $ref: "#/definitions/EligibilityResponse"
        201:
          description: "Created"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
      deprecated: false
definitions:
  AddressDTO:
    type: "object"
    properties:
      addressLine1:
        type: "string"
        example: "Flat B"
        description: "First line of the address"
      addressLine2:
        type: "string"
        example: "221 Baker Street"
        description: "Second line of the address"
      postcode:
        type: "string"
        example: "AA1 1AA"
        description: "The postcode of the address."
      townOrCity:
        type: "string"
        example: "London"
        description: "Town or city of the address"
    title: "AddressDTO"
    description: "The address object"
  EligibilityRequest:
    type: "object"
    properties:
      ctcAnnualIncomeThreshold:
        type: "number"
        example: 11000.0
        description: "The annual income threshold for Child Tax Credit"
      eligibleEndDate:
        type: "string"
        format: "date"
        example: "2019-02-01"
        description: "The end date for eligibility"
      eligibleStartDate:
        type: "string"
        format: "date"
        example: "2019-01-01"
        description: "The start date for eligibility"
      person:
        description: "Details of the person"
        $ref: "#/definitions/PersonDTO"
    title: "EligibilityRequest"
    description: "The eligibility request for Child Tax Credit"
  EligibilityResponse:
    type: "object"
    properties:
      eligibilityStatus:
        type: "string"
        example: "ELIGIBLE"
        description: "The eligibility status"
        enum:
        - "ELIGIBLE"
        - "INELIGIBLE"
        - "PENDING"
        - "NOMATCH"
      householdIdentifier:
        type: "string"
        example: "9d6049ee-d6e6-4d9b-ae3d-3b1bdf98682f"
        description: "The identifier for their household"
      numberOfChildrenUnderFour:
        type: "integer"
        format: "int32"
        example: 1
        description: "The number of children under 4 that the person has (which will\
          \ include the number of children under 1)"
      numberOfChildrenUnderOne:
        type: "integer"
        format: "int32"
        example: 1
        description: "The number of children under 1 that the person has"
    title: "EligibilityResponse"
    description: "The eligibility response from HMRC"
  PersonDTO:
    type: "object"
    properties:
      address:
        description: "The person's address"
        $ref: "#/definitions/AddressDTO"
      dateOfBirth:
        type: "string"
        format: "date"
        example: "1985-12-30"
        description: "The date of birth, in the format YYYY-MM-DD"
      firstName:
        type: "string"
        example: "Lisa"
        description: "The person's first name"
      lastName:
        type: "string"
        example: "Simpson"
        description: "The person's last name"
      nino:
        type: "string"
        example: "QQ123456C"
        description: "National Insurance number"
        pattern: "[a-zA-Z]{2}\\d{6}[a-dA-D]"
    title: "PersonDTO"
    description: "The person who we are checking eligibility for"